+++ CHANGELOG.md
@@ -11,15 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For releases pre-1.0, see the [GitHub Releases](https://github.com/JuliaQuantumControl/GRAPE.jl/releases).
 
+
 ## [Unreleased]
 
 * Changed: The minimum supported Julia version is now 1.10 (LTS)
 * Changed: The minimum supported versions of dependencies are now QuantumControl 0.11.5 and QuantumGradientGenerators 0.1.9. GRAPE does not work with QuantumGradientGenerators 0.1.8, which lacks the type-based `supports_inplace` trait of QuantumPropagators 0.9
 
+
 ## [v1.1.0] — 2026-06-20
 
 * Added: Support for state-dependent running costs [[#53], [#103], followup in [#105]]
 * Fixed: Bug in `gradient_method = :taylor` incorrectly accessing `pulsevals`, leading to incorrect gradients when the control problem contains more than one control [fixed as part of [#103]]
+
 
 ## [v1.0.0] — 2025-10-30